add PercentageStrategyTest, and fix interface

+++ src/Contracts/DiscountContract.php
@@ -9,7 +9,7 @@ interface DiscountContract
      * Make Discount
      *
      * @param int $basePrice
-     * @return int
+     * @return float
      */
-    public function make(int $basePrice) : int;
+    public function make(int $basePrice) : float;
 }
