+++ src/DiscountStrategy/PercentageStrategy.php
@@ -21,9 +21,9 @@ class PercentageStrategy implements DiscountContract
      * Make Discount
      *
      * @param int $basePrice
-     * @return int
+     * @return float
      */
-    public function make(int $basePrice): int
+    public function make(int $basePrice): float
     {
         return $basePrice / 100 * $this->sign;
     }
